refactor(LadderBooleanGenerator): 사다리를 놓는 여부를 결정하는 `generator`가 외부에서 주어진 값을 가공할 수 있도록 수정
