+++ src/main/java/LadderMain.java
@@ -1,6 +1,7 @@
 import game.LadderGame;
 import generator.BooleanGenerator;
-import generator.LadderRandomBooleanGenerator;
+import generator.LadderBooleanGenerator;
+import java.util.Random;
 import view.ConsoleReader;
 import view.InputView;
 import view.OutputView;
@@ -11,7 +12,8 @@ public class LadderMain {
         ConsoleReader consoleReader = new ConsoleReader(System.in);
         InputView inputView = new InputView(consoleReader);
         OutputView outputView = new OutputView();
-        BooleanGenerator generator = new LadderRandomBooleanGenerator();
+        Random random = new Random();
+        BooleanGenerator generator = new LadderBooleanGenerator(random::nextBoolean);
 
         LadderGame ladderGame = new LadderGame(inputView, outputView, generator);
         ladderGame.play();
